voice_agent/requirements.txt
vosk
sounddevice
httpx
python-dotenv
numpy

